feat(job):[TRI-338] Move job related AASWrapper classes to own package
