+++ irs-api/src/main/java/net/catenax/irs/aaswrapper/job/AASRecursiveJobHandler.java
@@ -7,7 +7,7 @@
 // See the LICENSE file(s) distributed with this work for
 // additional information regarding license terms.
 //
-package net.catenax.irs.aaswrapper;
+package net.catenax.irs.aaswrapper.job;
 
 import java.util.stream.Stream;
